improve animation player

+++ src/importer/r3d_importer_skeleton.c
@@ -76,7 +76,7 @@ static void upload_skeleton_bind_pose(R3D_Skeleton* skeleton)
 
     glGenTextures(1, &skeleton->texBindPose);
     glBindTexture(GL_TEXTURE_1D, skeleton->texBindPose);
-    glTexImage1D(GL_TEXTURE_1D, 0, GL_RGBA32F, 4 * skeleton->boneCount, 0, GL_RGBA, GL_FLOAT, finalBindPose);
+    glTexImage1D(GL_TEXTURE_1D, 0, GL_RGBA16F, 4 * skeleton->boneCount, 0, GL_RGBA, GL_FLOAT, finalBindPose);
     glTexParameteri(GL_TEXTURE_1D, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
     glTexParameteri(GL_TEXTURE_1D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
     glTexParameteri(GL_TEXTURE_1D, GL_TEXTURE_WRAP_S, GL_CLAMP_TO_EDGE);
